feat: change money int to float

+++ api/endpoints/money/model.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class MoneyRecord(BaseModel):
-    money: int
+    money: float
     proposal_id: int
     sponsor_addr: str
     transaction_hash: str
