+++ api/endpoints/withdrawal_list/model.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class WithdrawalRecord(BaseModel):
-    money: int
+    money: float
     proposal_id: int
     use_description: str
     transaction_hash: str
