+++ sql/table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS proposal(
     proposal_id INTEGER PRIMARY KEY,
     owner_addr CHAR(42) NOT NULL,
-    target_price INTEGER NOT NULL,
+    target_price DOUBLE PRECISION NOT NULL,
     project_description CHAR(1000) NOT NULL,
     start_time TIMESTAMP NOT NULL,
     project_name CHAR(100) NOT NULL,
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS image_list(
 CREATE TABLE IF NOT EXISTS money_list(
     id SERIAL PRIMARY KEY,
     proposal_id INTEGER NOT NULL,
-    money INTEGER NOT NULL,
+    money DOUBLE PRECISION NOT NULL,
     sponsor_addr CHAR(42) NOT NULL,
     transaction_hash CHAR(66) NOT NULL,
     input_time TIMESTAMP NOT NULL,
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS money_list(
 CREATE TABLE IF NOT EXISTS withdrawal_list(
     id SERIAL PRIMARY KEY,
     proposal_id INTEGER NOT NULL,
-    money INTEGER NOT NULL,
+    money DOUBLE PRECISION NOT NULL,
     use_description CHAR(1000) NOT NULL,
     transaction_hash CHAR(66) NOT NULL,
     output_time TIMESTAMP NOT NULL,
